requirements.txt
astroid==2.3.2
astropy==3.2.2
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
biopython==1.73
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.2
colorama==0.4.1
confuse==1.0.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
GPy==1.9.8
GPyOpt==1.2.5
htmlmin==0.1.12
importlib-metadata==0.23
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.21
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.0.1
lazy-object-proxy==1.4.2
llvmlite==0.30.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mkl-service==2.0.2
more-itertools==7.2.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numba==0.46.0
numpy==1.16.2
olefile==0.46
packaging==19.2
pandas==0.24.2
pandas-profiling==2.3.0
pandocfilters==1.4.2
paramz==0.9.5
parso==0.3.4
phik==0.9.8
pickleshare==0.7.5
Pillow==5.4.1
plotly==4.1.1
pluggy==0.13.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
py==1.8.0
pycparser==2.19
Pygments==2.3.1
pylint==2.4.3
pyltr==0.2.4
pyparsing==2.3.1
pyrsistent==0.14.11
pytest==5.2.1
pytest-pylint==0.14.1
python-dateutil==2.8.0
pytz==2018.9
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.0.0
qtconsole==4.4.3
retrying==1.3.3
scikit-learn==0.20.3
scikit-optimize==0.5.2
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
terminado==0.8.1
testpath==0.4.2
torch==1.0.1
torchvision==0.2.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
